Implement Ability To Drag Selection Actions Bar
