+++ dataweaver/apps/web/src/components/scopes/atlas/components/in_front_of_canvas/index.tsx
@@ -1,7 +1,7 @@
 import { Portal } from '~/components/primitives/portal';
 import { Controls } from './controls';
 import s from './index.module.scss';
-import { Selection } from './selection';
+import { Selection } from './selection/selection';
 import { Tools } from './tools';
 
 export const InFrontOfTheCanvas = () => {
